+++ functions/Inputs.php
@@ -1876,21 +1876,20 @@ function InputDivOnclick( $id, $input_html, $value, $input_ftitle )
 {
 	// @since 9.0 JS Sanitize string for legal variable name.
 	// @link https://stackoverflow.com/questions/12339942/sanitize-strings-for-legal-variable-names-in-php
-	$pattern = '/^(?![a-zA-Z_\x7f-\xff][a-zA-Z0-9_\x7f-\xff]*)$/';
+	$id_sanitized = preg_replace( '/^(?![a-zA-Z_\x7f-\xff][a-zA-Z0-9_\x7f-\xff]*)$/', '', $id );
 
-	$id_var_name_sanitized = preg_replace( $pattern, '', $id );
-
-	$script = '<script>var html' . $id_var_name_sanitized . '=' . json_encode( $input_html ).';</script>';
+	$script = '<script>var html' . $id_sanitized . '=' . json_encode( $input_html ).';</script>';
 
 	$value = $value == '' ? '-' : $value;
 
-	$onfocus_js = 'addHTML(html' . $id_var_name_sanitized . ',"div' . $id_var_name_sanitized . '",true);
-		$("#' . $id_var_name_sanitized . '").focus();
-		$("#div' . $id_var_name_sanitized . '").click();';
+	$onfocus_js = 'addHTML(html' . $id_sanitized . ',\'div' . $id_sanitized . '\',true);
+		$(\'#' . $id_sanitized . '\').focus(); $(\'#div' . $id_sanitized . '\').click();';
 
-	$div_onclick = '<div id="div' . $id_var_name_sanitized . '">
-		<div class="onclick" tabindex="0" onfocus="' . AttrEscape( $onfocus_js ) . '">' .
-		( mb_strpos( $value, '<div' ) === 0 ?
+	$div_onclick = '<div id="div' . $id_sanitized . '">
+		<div class="onclick" tabindex="0" onfocus="' .
+		// Do not not convert single quotes to gain a few bytes
+		htmlspecialchars( $onfocus_js, ENT_COMPAT, null, false ) . '">' .
+		( mb_stripos( $value, '<div' ) === 0 ?
 			'<div class="underline-dots">' . $value . '</div>' :
 			'<span class="underline-dots">' . $value . '</span>' ) .
 		$input_ftitle . '</div></div>';